# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: M15-AUDIT API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ledger/closing-status/{clientId}/{year}:
    get:
      operationId: getClosingStatus
      tags: [closing]
      summary: Get the fiscal year closing status for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema: { type: integer }
        - name: year
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Closing status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClosingStatus"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /ledger/close-period/{clientId}/{year}:
    post:
      operationId: closePeriod
      tags: [closing]
      summary: Lock a fiscal year (Clôturer définitivement)
      description: >
        Runs the full 4-step SYSCOHADA year-end closing routine:
        (1) generate and auto-validate pending depreciation dotations and financial installments,
        (2) compute net result from Class 6/7 balances and post the clearing entry to account 1301 (bénéfice) or 1309 (perte),
        (3) lock the period so no further ledger entries are accepted for this client/year,
        (4) generate the Journal des À-nouveaux carrying forward Class 1-5 balances to year+1.
        Restricted to expert_comptable.
      parameters:
        - name: clientId
          in: path
          required: true
          schema: { type: integer }
        - name: year
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Period closed and locked successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClosePeriodResult"
        "403":
          description: Period already locked or insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /tax/dsf/{clientId}/{year}:
    get:
      operationId: getDsf
      tags: [tax]
      summary: Compute the full DSF / Liasse Fiscale SYSCOHADA Révisé (Bilan, Compte de Résultat, TFT) live from the validated ledger (M24)
      parameters:
        - name: clientId
          in: path
          required: true
          schema: { type: integer }
        - name: year
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Computed DSF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DsfResult"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    FiscalYearClosingStatus:
      type: string
      enum: [OPEN, LOCKED]
    ClosingStatus:
      type: object
      required: [clientId, year, status, openingBalanceGenerated]
      properties:
        clientId: { type: integer }
        year: { type: integer }
        status:
          $ref: "#/components/schemas/FiscalYearClosingStatus"
        netResult: { type: ["integer", "null"] }
        netResultAccount: { type: ["string", "null"] }
        openingBalanceGenerated: { type: boolean }
        lockedAt: { type: ["string", "null"], format: date-time }
        lockedByName: { type: ["string", "null"] }
    ClosePeriodResult:
      type: object
      required: [clientId, year, step1, step2, lockedAt, step4]
      properties:
        clientId: { type: integer }
        year: { type: integer }
        step1:
          type: object
          required:
            [depreciationEntriesGenerated, depreciationEntriesSkipped, financeEntriesGenerated, financeEntriesSkipped]
          properties:
            depreciationEntriesGenerated: { type: integer }
            depreciationEntriesSkipped: { type: integer }
            financeEntriesGenerated: { type: integer }
            financeEntriesSkipped: { type: integer }
        step2:
          type: object
          required: [totalClass6Debits, totalClass7Credits, netResult, resultAccount, closingTransactionId]
          properties:
            totalClass6Debits: { type: integer }
            totalClass7Credits: { type: integer }
            netResult: { type: integer }
            resultAccount: { type: string }
            closingTransactionId: { type: integer }
        lockedAt: { type: string, format: date-time }
        step4:
          type: object
          required: [accountsCarriedForward]
          properties:
            accountsCarriedForward: { type: integer }
            openingTransactionId: { type: ["integer", "null"] }

    # -------------------------------------------------------------------------
    # Plan Comptable (SYSCOHADA chart of accounts) — read-only lookup used by
    # account-autocomplete inputs (e.g. Balance d'Entrée grid below).
    # -------------------------------------------------------------------------
...
    DsfBilanActifLine:
      type: object
      required: [lineCode, label, isSubtotal, isSectionHeader, brut, amortissements, netN]
      properties:
        lineCode: { type: string }
        label: { type: string }
        isSubtotal: { type: boolean }
        isSectionHeader: { type: boolean }
        brut: { type: integer }
        amortissements: { type: integer }
        netN: { type: integer }
    DsfBilanPassifLine:
      type: object
      required: [lineCode, label, isSubtotal, isSectionHeader, montantN]
      properties:
        lineCode: { type: string }
        label: { type: string }
        isSubtotal: { type: boolean }
        isSectionHeader: { type: boolean }
        montantN: { type: integer }
    DsfCompteResultatLine:
      type: object
      required: [lineCode, label, produits, charges, solde, isIntermediate, isSectionHeader]
      properties:
        lineCode: { type: string }
        label: { type: string }
        produits: { type: integer }
        charges: { type: integer }
        solde: { type: integer }
        isIntermediate: { type: boolean }
        isSectionHeader: { type: boolean }
    DsfTftLine:
      type: object
      required: [lineCode, label, montantN, isSubtotal, isSectionHeader]
      properties:
        lineCode: { type: string }
        label: { type: string }
        montantN: { type: integer }
        isSubtotal: { type: boolean }
        isSectionHeader: { type: boolean }
    DsfResult:
      type: object
      required: [bilanActif, bilanPassif, compteResultat, tft, totalBilanActif, totalBilanPassif, balanceEquilibre, bilanEquilibre, totalDebits, totalCredits]
      properties:
        bilanActif:
          type: array
          items: { $ref: "#/components/schemas/DsfBilanActifLine" }
        bilanPassif:
          type: array
          items: { $ref: "#/components/schemas/DsfBilanPassifLine" }
        compteResultat:
          type: array
          items: { $ref: "#/components/schemas/DsfCompteResultatLine" }
        tft:
          type: array
          items: { $ref: "#/components/schemas/DsfTftLine" }
        totalBilanActif: { type: integer }
        totalBilanPassif: { type: integer }
        balanceEquilibre: { type: boolean }
        bilanEquilibre: { type: boolean }
        totalDebits: { type: integer }
        totalCredits: { type: integer }

    # -------------------------------------------------------------------
    # Module M27 (Scoring Financier & Évaluation d'Entreprise)
    # -------------------------------------------------------------------
...